5197: Barrel files refactor: meta/expressionEvaluator
